requirements.txt
numpy
opencv-contrib-python
Pillow
tensorflow
pandas
google-api-python-client
gspread-dataframe
pprintpp
oauth2client
gspread
tk